add: split to_ts func
