+++ src/to_ts_rust/to_ts/func/mod.rs
@@ -1,0 +1,1 @@
+pub mod func;
