+++ src/to_ts_rust/to_ts/mod.rs
@@ -2,3 +2,4 @@ pub mod to_ts;
 mod common_methods;
 mod variable_declaration;
 mod expr_stmt;
+mod func;
